% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute_info.R
\name{attribute_info}
\alias{attribute_info}
\title{Information about attribute}
\usage{
attribute_info(attrId, lang = c("pl", "en"), ...)
}
\arguments{
\item{attrId}{A single attribute id.Use \code{\link{get_attributes}} to find more info.}

\item{lang}{A language of returned data, "pl" (default), "en"}

\item{...}{Other arguments passed on to \code{\link[httr]{GET}}. For example
a proxy parameters, see details.}
}
\value{
A named list.
}
\description{
Retrieve information about attribute.
}
\details{
To use a proxy to connect, a \code{\link[httr]{use_proxy}} can be
  passed to \code{\link[httr]{GET}}. For example
  \code{get_request(id, filters,
  config = httr::use_proxy(url, port, username, password))}.
}
\examples{
 \donttest{
   attribute <- attribute_info("1")
}
}
\keyword{info}
\keyword{variables}
