% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdl.R
\docType{package}
\name{bdl}
\alias{bdl}
\alias{bdl-package}
\title{bdl: Interface and Tools for BDL API}
\description{
The 'bdl' package is an interface to Local Data Bank(Bank Danych Lokalnych - bdl) API 
    with set of useful tools like quick plotting and map generating using data from data bank.
}
\author{
\strong{Maintainer}: Krzysztof Kania \email{K.Kania@stat.gov.pl}

}
\keyword{internal}
