% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_levels.R
\name{get_levels}
\alias{get_levels}
\title{Get all levels}
\usage{
get_levels(sort = c("id", "-id", "name", "-name"), lang = c("pl", "en"), ...)
}
\arguments{
\item{sort}{A type of sorting, "id" (default), "-id", "name", "-name"}

\item{lang}{A language of returned data, "pl" (default), "en"}

\item{...}{Other arguments passed on to \code{\link[httr]{GET}}. For example
a proxy parameters, see details.}
}
\value{
A dataset as a tibble.
}
\description{
Retrieve all levels with information.
}
\details{
To use a proxy to connect, a \code{\link[httr]{use_proxy}} can be
  passed to \code{\link[httr]{GET}}. For example
  \code{get_request(id, filters,
  config = httr::use_proxy(url, port, username, password))}.
}
\examples{
   # get_levels()
}
\keyword{levels}
\keyword{utilities}
