% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{level1_map}
\alias{level1_map}
\title{Spatial data map of NUTS level 1 units of Poland 2016}
\format{A `sf` map with 7 rows and 3 variables:
\describe{
  \item{name}{name of unit}
  \item{id}{unit's NUTS id code}
  \item{geometry}{Shp coordinates}
}}
\usage{
level1_map
}
\description{
A `sf` map from `shp` with NUTS level 1 units of Poland 2010-2018 with its
name and NUTS id's.
}
\keyword{datasets}
