% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{level6_map}
\alias{level6_map}
\title{Spatial data map of NUTS level 6 units of Poland}
\format{A `sf` map with 2478 rows and 3 variables:
\describe{
  \item{name}{name of unit}
  \item{id}{unit's NUTS id code}
  \item{geometry}{Shp coordinates}
}}
\usage{
level6_map
}
\description{
A `sf` map from `shp` with NUTS level 6 units of Poland 2016 with its
name and NUTS id's.
}
\keyword{datasets}
