% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{level6_map_changes}
\alias{level6_map_changes}
\title{Spatial data frame with level 6 units that area has been changed}
\format{A `sf` map with 1 rows and 3 variables:
\describe{
  \item{name}{name of unit}
  \item{id}{unit's NUTS id code}
  \item{year}{year when change applies}
  \item{geometry}{Shp coordinates}
}}
\usage{
level6_map_changes
}
\description{
A `sf` data frame with spatial data for units that area has been changed
without changing its NUTS id. You should replace corresponding rows in 
`level6_map_units` with updated geometry from this data frame if 
data's year is higher or eagual to given in `year` column.
}
\keyword{datasets}
