% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{level6_map_units}
\alias{level6_map_units}
\title{Spatial data map of NUTS level 6 units of Poland}
\format{A `sf` map with 2511 rows and 3 variables:
\describe{
  \item{name}{name of unit}
  \item{id}{unit's NUTS id code}
  \item{geometry}{Shp coordinates}
}}
\usage{
level6_map_units
}
\description{
A `sf` map from `shp` with all unique NUTS level 6 units of Poland 
from 2010-2018 with itsname and NUTS id's. They should bee filtered
by inner join with corresponding NUTS ids from given year.
}
\keyword{datasets}
