% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subject_info.R
\name{subject_info}
\alias{subject_info}
\title{Information about subject}
\usage{
subject_info(subjectId, lang = c("pl", "en"), ...)
}
\arguments{
\item{subjectId}{A subject id code. If not specified returns all top level subjects.
Use \code{\link{search_subjects}} or \code{\link{get_subjects}} to find subject codes.}

\item{lang}{A language of returned data, "pl" (default), "en"}

\item{...}{Other arguments passed on to \code{\link[httr]{GET}}. For example
a proxy parameters, see details.}
}
\value{
A named list.
}
\description{
Retrieve information about subject
}
\details{
To use a proxy to connect, a \code{\link[httr]{use_proxy}} can be
  passed to \code{\link[httr]{GET}}. For example
  \code{get_request(id, filters,
  config = httr::use_proxy(url, port, username, password))}.
}
\examples{
 \donttest{
   subject_info("G7")
}
}
\keyword{info}
\keyword{subjects}
