% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.bdl}
\alias{summary.bdl}
\title{Summarize bdl data frame}
\usage{
\method{summary}{bdl}(object, ...)
}
\arguments{
\item{object}{bdl data frame to summarise}

\item{...}{other arguments ignored (for compatibility with generic)}
}
\description{
Prints brief summary with basic statistical functions like mean, standard deviation, variance,
         min and max for bdl data frame.
}
\examples{
# df <- get_data_by_variable(varId = "3643")
# summary(df)
}
