% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_info.R
\name{unit_info}
\alias{unit_info}
\title{Information about unit}
\usage{
unit_info(unitId, lang = c("pl", "en"), ...)
}
\arguments{
\item{unitId}{A 12 character NUTS id code of interested unit. Use \code{\link{search_units}} or
\code{\link{get_units}} to find unit id code.}

\item{lang}{A language of returned data, "pl" (default), "en"}

\item{...}{Other arguments passed on to \code{\link[httr]{GET}}. For example
a proxy parameters, see details.}
}
\value{
A named list.
}
\description{
Retrieve information about unit
}
\details{
To use a proxy to connect, a \code{\link[httr]{use_proxy}} can be
  passed to \code{\link[httr]{GET}}. For example
  \code{get_request(id, filters,
  config = httr::use_proxy(url, port, username, password))}.
}
\examples{
   # unit_info("030210106062")
}
\keyword{info}
\keyword{units}
