% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_locality_info.R
\name{unit_locality_info}
\alias{unit_locality_info}
\title{Information about unit locality}
\usage{
unit_locality_info(unitId, lang = c("pl", "en"), ...)
}
\arguments{
\item{unitId}{A 12 character NUTS id with 7 characters locality individual id, separated by dash.
Use \code{\link{search_unit_localities}} or \code{\link{get_unit_localities}} to find unit id code.}

\item{lang}{A language of returned data, "pl" (default), "en"}

\item{...}{Other arguments passed on to \code{\link[httr]{GET}}. For example
a proxy parameters, see details.}
}
\value{
A named list.
}
\description{
Retrieve information about unit locality
}
\details{
To use a proxy to connect, a \code{\link[httr]{use_proxy}} can be
  passed to \code{\link[httr]{GET}}. For example
  \code{get_request(id, filters,
  config = httr::use_proxy(url, port, username, password))}.
}
\examples{
   # unit_locality_info("030210106062-0189782")
}
\keyword{info}
\keyword{localities}
\keyword{units}
