% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_info.R
\name{variable_info}
\alias{variable_info}
\title{Information about variable}
\usage{
variable_info(varId, lang = c("pl", "en"), ...)
}
\arguments{
\item{varId}{A vector of variable Id's.Use \code{\link{search_variables}} or
\code{\link{get_variables}} to find variable id code.}

\item{lang}{A language of returned data, "pl" (default), "en"}

\item{...}{Other arguments passed on to \code{\link[httr]{GET}}. For example
a proxy parameters, see details.}
}
\value{
A named list.
}
\description{
Retrieve information about variable.
}
\details{
To use a proxy to connect, a \code{\link[httr]{use_proxy}} can be
  passed to \code{\link[httr]{GET}}. For example
  \code{get_request(id, filters,
  config = httr::use_proxy(url, port, username, password))}.
}
\examples{
   # variable_info("420")
}
\keyword{info}
\keyword{variables}
