% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdl.R
\docType{package}
\name{bdl}
\alias{bdl}
\alias{bdl-package}
\title{bdl: Interface and Tools for BDL API}
\description{
The 'bdl' package is an interface to Local Data Bank(Bank Danych Lokalnych - bdl) API 
    with set of useful tools like quick plotting and map generating using data from data bank.
}
\section{Package options}{

\describe{
\item{`bdl.api_private_key`}{String with BDL API key which you can get
\url{https://bdl.stat.gov.pl/api/v1/client?theme=Default} 
Example: \code{options(bdl.api_private_key = "3sd1dea1-f89d-488c-c441-08d64fbd20e9")}}

}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kaniakrzysztof/bdl}
  \item Report bugs at \url{https://github.com/kaniakrzysztof/bdl/issues}
}

}
\author{
\strong{Maintainer}: Krzysztof Kania \email{K.Kania@stat.gov.pl}

}
