% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdl.R
\docType{package}
\name{bdl}
\alias{bdl}
\alias{bdl-package}
\title{bdl: Interface and Tools for 'BDL' API}
\description{
Interface to Local Data Bank ('Bank Danych Lokalnych' - 'bdl') API 
    <https://api.stat.gov.pl/Home/BdlApi?lang=en> with set of useful tools like 
    quick plotting and map generating using data from bank.
}
\section{Package options}{

\describe{
\item{`bdl.api_private_key`}{String with BDL API key which you can get
\url{https://bdl.stat.gov.pl/api/v1/client?theme=Default} 
Example: \code{options(bdl.api_private_key = "11111111-2222-3333-4444-555555555555")}}

}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/statisticspoland/R_Package_to_API_BDL}
  \item Report bugs at \url{https://github.com/statisticspoland/R_Package_to_API_BDL/issues}
}

}
\author{
\strong{Maintainer}: Krzysztof Kania \email{K.Kania@stat.gov.pl}

Other contributors:
\itemize{
  \item Statistics Poland [copyright holder, funder]
}

}
